Calc service

+++ math-functions-lib/src/main/java/org/example/Adder.java
@@ -2,11 +2,11 @@ package main.java.org.example;
 
 public class Adder {
 
-    public int add( int i, int j ) {
+    public double add( double i, double j ) {
         return i + j;
     }
 
-    public int add2( int i, int j ) {
+    public double add2( double i, double j ) {
         return i + j;
     }
